+++ Changes.md
@@ -9,6 +9,9 @@ Version 4.0.0.1 - TBR
   4. Add the ability to specify icon font classes for the one section link.
   5. Implications of MDL-73679.
   6. Implications of MDL-73683.
+  7. Implications of MDL-73679.
+
+Note: Content -> Section -> Section Content -> CMList -> CMItem -> CM -> Activity.  CT Wraps at CMList level.
 
 Version 3.11.0.5 - 15/11/21
 -----------------------------